:hammer_and_wrench:  refactored services directory structure
