+++ services/shop/api/shopify/utils/normalizeProduct.ts
@@ -1,5 +1,5 @@
 import { Product as ShopifyProduct, ImageEdge } from '../schema.d';
-import { Product } from '@common/types/product';
+import { Product } from '@shop/types/product';
 
 const normalizeImages = ({ edges }: { edges: Array<ImageEdge> }) =>
   edges.map(({ node: { originalSrc: url, ...more } }) => ({
