+++ services/shop/api/woocommerce/next.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  accessKeyID: process.env.SHOP_ACCESS_KEY_ID,
+  accessSecretKey: process.env.SHOP_ACCESS_SECRET_KEY,
+  isWoo: true,
+};
