--- services/shopify/next.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-    test: 'shopify-settings',
-};
